Remove unused generic parameters from ClientSettings
The menu settings interfaces were only using the generic parameters for
specifying the FormApplication type to be sued for the menu. Removes
them since that functionality has been superceded by the ConstructorOf
type.

Unconstrains the data type in the FormApplication used for
ClientSettings.registerMenu to avoid errors on certain confugurations of
custom settings forms.