Use an alternate approach to defining types for ClientSettings
Determines the type to use for ClientSettings.register by checking if it
is defined in ClientSettings.Values. If it is, pass the type as a type
param to ClientSettings.PartialSetting<T> and type check according to
that. If it is not defined in Values, derive the type param for
PartialSetting based on the supplied data object.

Updates PartialSetting to be a little more robust about what type is
passed in. For example, PartialSetting<boolean> has a type of
ConstructorOf<Boolean> for the type field.

Adds testing to verify that augmentation of ClientSettings.Values
results in an authoritative typing of the fields.